debugging external port issue

+++ HEAR_lib/include/HEAR_library/Library.hpp
@@ -30,10 +30,10 @@ namespace HEAR
 
 class Library {
     public :
-    static Block* createBlock(BLOCK_ID b_type, int b_uid, float _dt = 0, TYPE d_type=TYPE::NA);
+    static Block* createBlock(BLOCK_ID b_type, int b_uid, double _dt = 0, TYPE d_type=TYPE::NA);
 };
 
-Block* Library::createBlock(BLOCK_ID b_type, int b_uid, float _dt, TYPE d_type){
+Block* Library::createBlock(BLOCK_ID b_type, int b_uid, double _dt, TYPE d_type){
     Block* blk;
     switch (b_type)
     {
